Refactor booking client integration
- Moved booking-related logic to a dedicated `client` package.
